Bug fix
Fix the bug for the sensor without the latest data but still showing in the button panel.
Everytime when it needs to change the value in the button, it will check with the timestamp now.
Several place when the sensor doesn't have the latest data, it will call noLatestSensor function to change all the button into grey color.
For those sensor doesn't have the latest value but still have historical data available, it can still be shown on the chart.